app/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.54
psycopg2-binary==2.9.10
fastapi-jwt-auth==0.5.0
pydantic==1.10.19
passlib==1.7.4
bcrypt==3.2.0
PyJWT==1.7.1
python-dotenv==1.0.1
reportlab==4.2.5
python-multipart
boto3
cryptography
python-dotenv
pydantic[email]
requests
matplotlib
pandas==2.2.3
redis
alembic==1.14.0

